refactor: Add gccrs mod and Wrapper type
